+++ src/daf/guild/__init__.py
@@ -1,0 +1,6 @@
+"""
+Subpackage containing guild related code.
+"""
+from .guilduser import *
+from .autoguild import *
+from .status import *
